+++ relay-dynamic-config/src/feature.rs
@@ -100,7 +100,7 @@ pub enum Feature {
     /// Enable the upload endpoint for attachments.
     #[serde(rename = "projects:relay-upload-endpoint")]
     UploadEndpoint,
-    /// Upload non-prosperodmp playstation attachments via the upload-endpoint.
+    /// Upload non-prosperodmp playstation attachments via the upload endpoint.
     #[serde(rename = "projects:relay-playstation-uploads")]
     PlaystationUploads,
     /// Add a random trace ID to events that lack one.
@@ -110,6 +110,9 @@ pub enum Feature {
     /// processor. Only enable for organizations with sufficient attachment quota.
     #[serde(rename = "organizations:relay-unreal-endpoint-expansion")]
     UnrealEndpointExpansion,
+    /// Upload minidump attachments via the upload endpoint.
+    #[serde(rename = "projects:relay-minidump-attachment-uploads")]
+    MinidumpAttachmentUploads,
 
     /// Enables OTLP spans to use the Span V2 processing pipeline in Relay.
     ///