+++ relay-server/src/endpoints/common.rs
@@ -1,21 +1,32 @@
 //! Common facilities for ingesting events through store-like endpoints.
+use std::io;
 
 use axum::http::{StatusCode, header};
 use axum::response::IntoResponse;
-use relay_config::RelayMode;
+use bytes::Bytes;
+use chrono::Utc;
+use futures::TryStreamExt;
+use futures::stream::BoxStream;
+use multer::Field;
+use relay_config::{Config, RelayMode};
 use relay_event_schema::protocol::{EventId, EventType};
-use relay_quotas::RateLimits;
+use relay_quotas::{RateLimits, Scoping};
 use relay_statsd::metric;
+use relay_system::Addr;
 use serde::Deserialize;
 
-use crate::envelope::{AttachmentType, Envelope, EnvelopeError, Item, ItemType, Items};
+use crate::envelope::{
+    AttachmentPlaceholder, AttachmentType, ContentType, Envelope, EnvelopeError, Item, ItemType,
+    Items,
+};
 use crate::managed::{Managed, Rejected};
 use crate::service::ServiceState;
 use crate::services::buffer::{ProjectKeyPair, PushError};
 use crate::services::outcome::{DiscardItemType, DiscardReason, Outcome};
 use crate::services::processor::{BucketSource, MetricData, ProcessMetrics};
+use crate::services::upload::{Create, Stream, Upload};
 use crate::statsd::{RelayCounters, RelayDistributions};
-use crate::utils::{self, ApiErrorResponse, FormDataIter};
+use crate::utils::{self, ApiErrorResponse, BoundedStream, FormDataIter, MeteredStream};
 
 #[derive(Clone, Copy, Debug, thiserror::Error)]
 #[error("the service is overloaded")]
@@ -139,7 +150,10 @@ impl IntoResponse for BadStoreRequest {
                 // now executed asynchronously in `EnvelopeProcessor`.
                 (StatusCode::FORBIDDEN, body).into_response()
             }
-            BadStoreRequest::Overflow(_) => (StatusCode::PAYLOAD_TOO_LARGE, body).into_response(),
+            BadStoreRequest::Overflow(_)
+            | BadStoreRequest::InvalidMultipart(multer::Error::StreamSizeExceeded { .. }) => {
+                (StatusCode::PAYLOAD_TOO_LARGE, body).into_response()
+            }
             _ => {
                 // In all other cases, we indicate a generic bad request to the client and render
                 // the cause. This was likely the client's fault.
@@ -442,6 +456,68 @@ fn emit_envelope_metrics(envelope: &Envelope) {
             sdk = client_name,
         );
     }
+}
+
+/// Uploads the content of `field` to the objectstore and returns an [Item] with an
+/// [AttachmentPlaceholder] as payload.
+///
+/// Returns `None` if uploading fails due to any reason.
+pub async fn upload_to_objectstore(
+    field: Field<'static>,
+    mut item: Item,
+    config: &Config,
+    scoping: Scoping,
+    upload: &Addr<Upload>,
+    referrer: &'static str,
+) -> Option<Item> {
+    let content_type = field.content_type().map(ToString::to_string);
+    let stream: BoxStream<'static, io::Result<Bytes>> = Box::pin(field.map_err(io::Error::other));
+    let stream = MeteredStream::new(stream, referrer);
+    let stream = BoundedStream::new(stream, 1, config.max_upload_size());
+    let byte_counter = stream.byte_counter();
+
+    let location = upload
+        .send(Create {
+            scoping,
+            length: None,
+        })
+        .await
+        .ok()?
+        .ok()?;
+
+    let result = upload
+        .send(Stream {
+            received: Utc::now(),
+            scoping,
+            location,
+            stream,
+        })
+        .await
+        .ok()?;
+
+    let location = result
+        .inspect_err(|e| {
+            relay_log::warn!(
+                error = e as &dyn std::error::Error,
+                referrer = referrer,
+                organization_id = scoping.organization_id.value(),
+                project_id = scoping.project_id.value(),
+                bytes_uploaded = byte_counter.get(),
+                "object store upload failed",
+            );
+        })
+        .ok()?;
+    let location = location.into_header_value().ok()?;
+    let location = location.to_str().ok()?;
+    let payload = serde_json::to_vec(&AttachmentPlaceholder {
+        location,
+        content_type,
+    })
+    .ok()?;
+
+    item.set_payload(ContentType::AttachmentRef, payload);
+    item.set_attachment_length(byte_counter.get());
+    Some(item)
 }
 
 #[derive(Debug)]